skip nvim install on non 86 target

+++ install.sh
@@ -18,11 +18,17 @@ done
 # TODO: Allow to switch from http checkout to checkout over ssh e.g. git remote add rw  git@github.com:keesj/my-lazy-vim.git
 # TODO: Add wget syntax
 mkdir -p opt
-rm -rf ~/opt/nvim-linux
-wget -qO- https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz | tar zxvf - -C ~/opt
-if ! grep -q '${HOME}/bin' ~/.bashrc; then
-	echo 'export PATH="${HOME}/bin:${PATH}"' >>~/.bashrc
-	echo 'alias vim="nvim"' >>~/.bashrc
+if uname -p | grep x86_64 2>&1 >/dev/null; then
+	rm -rf ~/opt/nvim-linux
+	wget -qO- https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz | tar zxvf - -C ~/opt
+	if ! grep -q '${HOME}/bin' ~/.bashrc; then
+		echo 'export PATH="${HOME}/bin:${PATH}"' >>~/.bashrc
+		echo 'alias vim="nvim"' >>~/.bashrc
+	fi
+else
+	echo "Skip nvim install"
+	echo https://github.com/neovim/neovim
+	sleep 5
 fi
 if [ ! -d bin ]; then
 	mkdir bin
